refactor(realtime-api): co-locate tests with source
Move apps/realtime-api/tests/* into src/**/__tests__/ to match the
web app convention; exclude __tests__ from the build tsconfig so
declaration emit stays clean; update lint scope and CLAUDE.md.
