+++ apps/realtime-api/tsconfig.json
@@ -9,5 +9,6 @@
     "lib": ["ES2020"],
     "types": ["node", "aws-lambda"]
   },
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": ["src/**/__tests__/**", "src/**/*.test.ts"]
 }
